moved test into dedicated file
