+++ tests/PHPStan/Analyser/data/math.php
@@ -109,20 +109,4 @@ class Foo
 		assertType('int', $i - $j);
 	}
 
-	public function unionSum(): void
-	{
-		$a = [];
-
-		if (rand(0,1) === 0) {
-			$a = ['a' => 5];
-		}
-
-		$b = [];
-
-		if (rand(0,1) === 0) {
-			$b = ['b' => 6];
-		}
-
-		assertType("array()|array(?'b' => 6, ?'a' => 5)", $a + $b);
-	}
 }
